Turn StringToTime into a static utility class
There is really no need for StringToTime to be used as an object
instance, especially since java Date is inferior to Joda DateTime.
Therefore, I've altered the class to preserve the parsing logic that
the original author wrote, but to function as a static util class and
either return a timestamp or a joda DateTime object. This greatly
simplifies the class.

Right now, the parsing still uses java Date under the hood, but that
may be changed in a future commit

Change-Id: Ice0dc3287117b0017dfea7c6b18b4850f048cdf1